requirements.txt
jira
sty
gitpython
ruamel.yaml

